+++ app/src/main/java/net/ozero/githubapp/usecase/repos/ObserveRepoByIdUseCase.kt
@@ -9,11 +9,11 @@ import net.ozero.githubapp.entity.Repo
 import net.ozero.githubapp.usecase.ObservableParamsUseCase
 import net.ozero.githubapp.usecase.ObservableUseCase
 
-interface ObserveRepoByIdUseCase : ObservableParamsUseCase<Repo, Int> {
+interface ObserveRepoByIdUseCase : ObservableParamsUseCase<Repo, Long> {
 
     class Impl(private val repoSource: RepoSource) : ObserveRepoByIdUseCase {
 
-        override suspend fun executeAsync(params: Int): Deferred<LiveData<Repo>> =
+        override suspend fun executeAsync(params: Long): Deferred<LiveData<Repo>> =
             CoroutineScope(Dispatchers.IO).async {
                 repoSource.observeById(params)
             }
